@@ -29,12 +29,14 @@
     <packaging>jar</packaging>
 
     <properties>
-        <dropwizard.version>3.0.13</dropwizard.version>
-        <jersey.version>2.35</jersey.version>
-        <logback.version>1.3.14</logback.version>
+        <dropwizard.version>4.0.13</dropwizard.version>
+        <jersey.version>3.0.17</jersey.version>
+        <logback.version>1.5.18</logback.version>
         <cxf.version>3.4.5</cxf.version>
-        <shiro.version>1.13.0</shiro.version>
-        <prometheus.version>0.12.0</prometheus.version>
+        <prometheus.version>0.16.0</prometheus.version>
+        <maven.compiler.source>11</maven.compiler.source>
+        <maven.compiler.target>11</maven.compiler.target>
+        <maven.compiler.release>11</maven.compiler.release>
         <docker.directory>src/main/docker</docker.directory>
         <timestamp>${maven.build.timestamp}</timestamp>
         <maven.build.timestamp.format>yyyy-MM-dd HH:mm:ss</maven.build.timestamp.format>
