@@ -19,12 +19,13 @@ package io.cassandrareaper.resources;
 
 import io.cassandrareaper.crypto.Cryptograph;
 
-import javax.ws.rs.GET;
-import javax.ws.rs.Path;
-import javax.ws.rs.PathParam;
-import javax.ws.rs.Produces;
-import javax.ws.rs.core.MediaType;
-import javax.ws.rs.core.Response;
+import jakarta.annotation.security.RolesAllowed;
+import jakarta.ws.rs.GET;
+import jakarta.ws.rs.Path;
+import jakarta.ws.rs.PathParam;
+import jakarta.ws.rs.Produces;
+import jakarta.ws.rs.core.MediaType;
+import jakarta.ws.rs.core.Response;
 
 @Path("/crypto")
 @Produces(MediaType.TEXT_PLAIN)
@@ -38,6 +39,7 @@ public final class CryptoResource {
 
   @GET
   @Path("/encrypt/{text}")
+  @RolesAllowed({"user", "operator"})
   public Response encrypt(@PathParam("text") String text) {
     return Response.ok().entity(cryptograph.encrypt(text)).build();
   }
