@@ -25,15 +25,15 @@ import io.cassandrareaper.storage.events.IEventsDao;
 import java.util.UUID;
 import java.util.concurrent.ScheduledExecutorService;
 
-import javax.servlet.http.HttpServletRequest;
-import javax.ws.rs.DefaultValue;
-import javax.ws.rs.GET;
-import javax.ws.rs.HeaderParam;
-import javax.ws.rs.Path;
-import javax.ws.rs.PathParam;
-import javax.ws.rs.Produces;
-import javax.ws.rs.core.Context;
-
+import jakarta.annotation.security.RolesAllowed;
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.ws.rs.DefaultValue;
+import jakarta.ws.rs.GET;
+import jakarta.ws.rs.HeaderParam;
+import jakarta.ws.rs.Path;
+import jakarta.ws.rs.PathParam;
+import jakarta.ws.rs.Produces;
+import jakarta.ws.rs.core.Context;
 import org.apache.hc.client5.http.impl.classic.CloseableHttpClient;
 import org.glassfish.jersey.media.sse.EventOutput;
 import org.glassfish.jersey.media.sse.SseFeature;
@@ -59,6 +59,7 @@ public final class DiagEventSseResource {
   @GET
   @Path("/{id}")
   @Produces(SseFeature.SERVER_SENT_EVENTS)
+  @RolesAllowed({"user", "operator"})
   public synchronized EventOutput listen(
       @Context HttpServletRequest request,
       @PathParam("id") String id,
