@@ -23,13 +23,6 @@ import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 
-import javax.annotation.Nullable;
-import javax.validation.Valid;
-import javax.validation.constraints.DecimalMin;
-import javax.validation.constraints.Max;
-import javax.validation.constraints.NotNull;
-import javax.ws.rs.DefaultValue;
-
 import com.datastax.oss.driver.api.core.addresstranslation.AddressTranslator;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import com.google.common.annotations.VisibleForTesting;
@@ -37,9 +30,14 @@ import io.dropwizard.cassandra.BasicCassandraFactory;
 import io.dropwizard.cassandra.CassandraFactory;
 import io.dropwizard.client.HttpClientConfiguration;
 import io.dropwizard.core.Configuration;
+import jakarta.annotation.Nullable;
+import jakarta.validation.Valid;
+import jakarta.validation.constraints.DecimalMin;
+import jakarta.validation.constraints.Max;
+import jakarta.validation.constraints.NotNull;
+import jakarta.ws.rs.DefaultValue;
 import org.apache.cassandra.repair.RepairParallelism;
 import org.hibernate.validator.constraints.NotEmpty;
-import org.secnod.dropwizard.shiro.ShiroConfiguration;
 
 public final class ReaperApplicationConfiguration extends Configuration {
 
@@ -120,7 +118,6 @@ public final class ReaperApplicationConfiguration extends Configuration {
   private Integer clusterTimeoutInDays;
 
   @JsonProperty private DatacenterAvailability datacenterAvailability;
-  @JsonProperty private AccessControlConfiguration accessControl;
   @JsonProperty private Integer repairThreadCount;
 
   /** If set to more than 0, defines how many days of run history should be kept. Default: 30 */
@@ -147,6 +144,8 @@ public final class ReaperApplicationConfiguration extends Configuration {
   @JsonProperty private Boolean scheduleRetryOnError;
 
   @JsonProperty private Duration scheduleRetryDelay;
+
+  @JsonProperty private AccessControlConfiguration accessControl;
 
   public HttpManagement getHttpManagement() {
     return httpManagement;
@@ -384,18 +383,6 @@ public final class ReaperApplicationConfiguration extends Configuration {
     this.datacenterAvailability = datacenterAvailability;
   }
 
-  public AccessControlConfiguration getAccessControl() {
-    return accessControl;
-  }
-
-  public void setAccessControl(AccessControlConfiguration accessControl) {
-    this.accessControl = accessControl;
-  }
-
-  public boolean isAccessControlEnabled() {
-    return getAccessControl() != null;
-  }
-
   public int getRepairThreadCount() {
     return repairThreadCount != null ? repairThreadCount : 1;
   }
@@ -513,6 +500,14 @@ public final class ReaperApplicationConfiguration extends Configuration {
 
   public void setScheduleRetryDelay(Duration scheduleRetryDelay) {
     this.scheduleRetryDelay = scheduleRetryDelay;
+  }
+
+  public AccessControlConfiguration getAccessControl() {
+    return accessControl;
+  }
+
+  public void setAccessControl(AccessControlConfiguration accessControl) {
+    this.accessControl = accessControl;
   }
 
   public enum DatacenterAvailability {
@@ -682,20 +677,6 @@ public final class ReaperApplicationConfiguration extends Configuration {
     }
   }
 
-  public static final class AccessControlConfiguration {
-
-    @JsonProperty private ShiroConfiguration shiro;
-    @JsonProperty private Duration sessionTimeout;
-
-    public ShiroConfiguration getShiroConfiguration() {
-      return shiro;
-    }
-
-    public Duration getSessionTimeout() {
-      return sessionTimeout != null ? sessionTimeout : Duration.ofMinutes(10);
-    }
-  }
-
   public static final class HttpManagement {
     @JsonProperty private Boolean enabled = false;
 
@@ -762,6 +743,87 @@ public final class ReaperApplicationConfiguration extends Configuration {
     @JsonProperty("mgmtApiMetricsPort")
     public void setMgmtApiMetricsPort(int mgmtApiMetricsPort) {
       this.mgmtApiMetricsPort = mgmtApiMetricsPort;
+    }
+  }
+
+  public static final class AccessControlConfiguration {
+    @JsonProperty private Boolean enabled = true;
+    @JsonProperty private Duration sessionTimeout;
+    @JsonProperty private JwtConfiguration jwt;
+    @JsonProperty private List<UserConfiguration> users = Collections.emptyList();
+
+    public Boolean isEnabled() {
+      return enabled;
+    }
+
+    public void setEnabled(Boolean enabled) {
+      this.enabled = enabled;
+    }
+
+    public Duration getSessionTimeout() {
+      return sessionTimeout != null ? sessionTimeout : Duration.ofMinutes(10);
+    }
+
+    public void setSessionTimeout(Duration sessionTimeout) {
+      this.sessionTimeout = sessionTimeout;
+    }
+
+    public JwtConfiguration getJwt() {
+      return jwt;
+    }
+
+    public void setJwt(JwtConfiguration jwt) {
+      this.jwt = jwt;
+    }
+
+    public List<UserConfiguration> getUsers() {
+      return users != null ? users : Collections.emptyList();
+    }
+
+    public void setUsers(List<UserConfiguration> users) {
+      this.users = users != null ? users : Collections.emptyList();
+    }
+  }
+
+  public static final class JwtConfiguration {
+    @JsonProperty private String secret;
+
+    public String getSecret() {
+      return secret;
+    }
+
+    public void setSecret(String secret) {
+      this.secret = secret;
+    }
+  }
+
+  public static final class UserConfiguration {
+    @JsonProperty private String username;
+    @JsonProperty private String password;
+    @JsonProperty private List<String> roles = Collections.emptyList();
+
+    public String getUsername() {
+      return username;
+    }
+
+    public void setUsername(String username) {
+      this.username = username;
+    }
+
+    public String getPassword() {
+      return password;
+    }
+
+    public void setPassword(String password) {
+      this.password = password;
+    }
+
+    public List<String> getRoles() {
+      return roles != null ? roles : Collections.emptyList();
+    }
+
+    public void setRoles(List<String> roles) {
+      this.roles = roles != null ? roles : Collections.emptyList();
     }
   }
 }
