@@ -19,10 +19,9 @@ package io.cassandrareaper.core;
 import io.cassandrareaper.validators.NullOrNotBlank;
 import io.cassandrareaper.validators.ValidEditableRepairSchedule;
 
-import javax.validation.constraints.Max;
-import javax.validation.constraints.Min;
-
 import com.fasterxml.jackson.annotation.JsonProperty;
+import jakarta.validation.constraints.Max;
+import jakarta.validation.constraints.Min;
 import org.apache.cassandra.repair.RepairParallelism;
 
 @ValidEditableRepairSchedule
