@@ -32,20 +32,20 @@ import java.util.UUID;
 import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.atomic.AtomicBoolean;
 
-import javax.ws.rs.DELETE;
-import javax.ws.rs.GET;
-import javax.ws.rs.POST;
-import javax.ws.rs.Path;
-import javax.ws.rs.PathParam;
-import javax.ws.rs.Produces;
-import javax.ws.rs.QueryParam;
-import javax.ws.rs.core.Context;
-import javax.ws.rs.core.MediaType;
-import javax.ws.rs.core.Response;
-import javax.ws.rs.core.UriInfo;
-
 import com.google.common.annotations.VisibleForTesting;
 import com.google.common.collect.ImmutableSet;
+import jakarta.annotation.security.RolesAllowed;
+import jakarta.ws.rs.DELETE;
+import jakarta.ws.rs.GET;
+import jakarta.ws.rs.POST;
+import jakarta.ws.rs.Path;
+import jakarta.ws.rs.PathParam;
+import jakarta.ws.rs.Produces;
+import jakarta.ws.rs.QueryParam;
+import jakarta.ws.rs.core.Context;
+import jakarta.ws.rs.core.MediaType;
+import jakarta.ws.rs.core.Response;
+import jakarta.ws.rs.core.UriInfo;
 import org.apache.hc.client5.http.impl.classic.CloseableHttpClient;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -72,6 +72,7 @@ public final class DiagEventSubscriptionResource {
   }
 
   @GET
+  @RolesAllowed({"user", "operator"})
   public Response getEventSubscriptionList(
       @QueryParam("clusterName") Optional<String> clusterName) {
     LOG.debug("get event subscriptions called %s", clusterName);
@@ -85,6 +86,7 @@ public final class DiagEventSubscriptionResource {
   }
 
   @POST
+  @RolesAllowed({"operator"})
   public Response addEventSubscription(
       @Context UriInfo uriInfo,
       @QueryParam("clusterName") String cluster,
@@ -140,6 +142,7 @@ public final class DiagEventSubscriptionResource {
 
   @GET
   @Path("/{id}")
+  @RolesAllowed({"user", "operator"})
   public Response getEventSubscription(@PathParam("id") UUID id) {
     LOG.debug("get subscription {}", id);
     try {
@@ -152,6 +155,7 @@ public final class DiagEventSubscriptionResource {
   @VisibleForTesting
   @GET
   @Path("/adhoc")
+  @RolesAllowed({"user", "operator"})
   public Response getActiveAdhocSubscriptions() {
     try {
       Collection<DiagEventSubscription> allSubs = eventsDao.getEventSubscriptions();
@@ -169,6 +173,7 @@ public final class DiagEventSubscriptionResource {
   @VisibleForTesting
   @GET
   @Path("/pollers")
+  @RolesAllowed({"user", "operator"})
   public Response getActivePollers() {
     try {
       if (DiagEventSubscriptionService.POLLERS_BY_NODE.isEmpty()) {
@@ -182,6 +187,7 @@ public final class DiagEventSubscriptionResource {
 
   @DELETE
   @Path("/{id}")
+  @RolesAllowed({"operator"})
   public Response deleteEventSubscription(@PathParam("id") UUID id) {
     LOG.debug("delete subscription {}", id);
     try {
