@@ -25,16 +25,16 @@ import io.cassandrareaper.service.MetricsService;
 import java.util.List;
 import java.util.Map;
 
-import javax.ws.rs.GET;
-import javax.ws.rs.Path;
-import javax.ws.rs.PathParam;
-import javax.ws.rs.Produces;
-import javax.ws.rs.core.Context;
-import javax.ws.rs.core.MediaType;
-import javax.ws.rs.core.Response;
-import javax.ws.rs.core.UriInfo;
-
 import com.google.common.base.Preconditions;
+import jakarta.annotation.security.RolesAllowed;
+import jakarta.ws.rs.GET;
+import jakarta.ws.rs.Path;
+import jakarta.ws.rs.PathParam;
+import jakarta.ws.rs.Produces;
+import jakarta.ws.rs.core.Context;
+import jakarta.ws.rs.core.MediaType;
+import jakarta.ws.rs.core.Response;
+import jakarta.ws.rs.core.UriInfo;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -62,6 +62,7 @@ public final class NodeStatsResource {
    */
   @GET
   @Path("/tpstats/{clusterName}/{host}")
+  @RolesAllowed({"user", "operator"})
   public Response getTpStats(
       @Context UriInfo uriInfo,
       @PathParam("clusterName") String clusterName,
@@ -88,6 +89,7 @@ public final class NodeStatsResource {
    */
   @GET
   @Path("/dropped/{clusterName}/{host}")
+  @RolesAllowed({"user", "operator"})
   public Response getDroppedMessages(
       @Context UriInfo uriInfo,
       @PathParam("clusterName") String clusterName,
@@ -114,6 +116,7 @@ public final class NodeStatsResource {
    */
   @GET
   @Path("/clientRequestLatencies/{clusterName}/{host}")
+  @RolesAllowed({"user", "operator"})
   public Response getClientRequestLatencies(
       @Context UriInfo uriInfo,
       @PathParam("clusterName") String clusterName,
@@ -140,6 +143,7 @@ public final class NodeStatsResource {
    */
   @GET
   @Path("/compactions/{clusterName}/{host}")
+  @RolesAllowed({"user", "operator"})
   public Response listCompactions(
       @Context UriInfo uriInfo,
       @PathParam("clusterName") String clusterName,
@@ -166,6 +170,7 @@ public final class NodeStatsResource {
    */
   @GET
   @Path("/tokens/{clusterName}/{host}")
+  @RolesAllowed({"user", "operator"})
   public Response listTokens(
       @Context UriInfo uriInfo,
       @PathParam("clusterName") String clusterName,
