@@ -18,13 +18,13 @@ package io.cassandrareaper.resources;
 
 import io.cassandrareaper.AppContext;
 
-import javax.ws.rs.GET;
-import javax.ws.rs.Path;
-import javax.ws.rs.Produces;
-import javax.ws.rs.core.MediaType;
-import javax.ws.rs.core.Response;
-
 import com.google.common.collect.ImmutableMap;
+import jakarta.annotation.security.RolesAllowed;
+import jakarta.ws.rs.GET;
+import jakarta.ws.rs.Path;
+import jakarta.ws.rs.Produces;
+import jakarta.ws.rs.core.MediaType;
+import jakarta.ws.rs.core.Response;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -48,6 +48,7 @@ public final class ReaperResource {
    */
   @GET
   @Path("/datacenterAvailability")
+  @RolesAllowed({"user", "operator"})
   public Response getDatacenterAvailability() {
     return Response.ok()
         .entity(
