@@ -23,10 +23,9 @@ import io.cassandrareaper.core.RepairUnit;
 import java.util.Collection;
 import java.util.UUID;
 
-import javax.annotation.Nullable;
-
 import com.fasterxml.jackson.annotation.JsonIgnore;
 import com.fasterxml.jackson.annotation.JsonProperty;
+import jakarta.annotation.Nullable;
 import org.apache.cassandra.repair.RepairParallelism;
 import org.apache.commons.lang3.time.DurationFormatUtils;
 import org.joda.time.DateTime;
